Add fragment graph
This add a iterator type object which carries a pointer to a modified
peptide and is able to cycle through every node of the graph of theoretical
fragments. It can be created for either b or y ions and any charge state. It could
also be modified later to add more ions. The calculations rely on a set of core
incrementation functions which could be modified to suit any ion type.