Adding function to determing site determining ions between signatures
These are peaks which are not present in one of the the peptide
permutations. This also adds a struct for saving scores and adding
ambiguity scores. I realized while coding the ambiguity scores that
I was having trouble keepiing track of all the scores. The easiest way
to fix this was to make a single ScoreContainer struct that was private
to the Ascore class which could  be sorted and had all the information
contained in one spot.