Remove full fragment from list of fragements
Originally, I was allowing each fragment graph to go all the way to
the end of a peptide to generate fragments. However, this makes no sense
since the last residue cannot create a true b or y fragment. This was
causing a slight discrepency between the scores of the two programs
and is fixed with this commit.